Update LocalDiskMapFileProvider and add new namespaces
Updated the copyright notice in `LocalDiskMapFileProvider.cs` to include "Modifications Copyright (C) 2022 Capnode AB". Added two new namespaces, `QuantConnect.Util` and `QuantConnect.Configuration`. Initialized a new `_dataProvider` field in the `LocalDiskMapFileProvider` constructor, which is set to an instance of `IDataProvider` retrieved from the `Composer` class using the type name specified in the configuration under the key "data-provider", with "DefaultDataProvider" as the default type name if no such configuration exists.