Add StopPrice to OrderViewModel and update BacktestView
Added a private field `_stopPrice` and a public property `StopPrice`
to the `OrderViewModel` class. Updated the class to initialize
`StopPrice` to null and set it based on the order type. Improved
readability in `BacktestView.xaml` by removing unnecessary line
breaks and combining attributes. Added a new `DataGridTextColumn`
for `Price` and changed its binding to `StopPrice` with the header
"Stop".